requirements.txt
asgiref==3.5.2
Django==4.1
sqlparse==0.4.2
djangorestframework==3.13.1
djoser==2.1.0